refactor: improve on code duplication
