+++ packages/lib/src/components/helpers/IssuerListContainer.tsx
@@ -102,7 +102,7 @@ class IssuerListContainer extends UIElement<IssuerListContainerProps> {
         return (
             <CoreProvider i18n={this.props.i18n} loadingContext={this.props.loadingContext}>
                 {this.props.issuers.length > 0 ? (
-                    <SRPanelProvider srPanel={this.props.modules.srPanel}>
+                    <SRPanelProvider srPanel={this.props?.modules?.srPanel}>
                         <IssuerList
                             ref={ref => {
                                 this.componentRef = ref;
